Add files via upload
Trouble in with the temp_df_sophia. I have been trying to trouble shoot the issue. I think the issue might be that between char_a and char_b on sophia screentime df there are no similar values. The df keeps coming up blank, which is wrong. Perhaps the lengths of the columns are different?